ss11

+++ src/ss10/product_cart/src/main/java/com/cart/repository/IProductRepository.java
@@ -3,5 +3,5 @@ package com.cart.repository;
 import com.cart.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProductRepository extends JpaRepository<Product, Long> {
+public interface IProductRepository extends JpaRepository<Product, Integer> {
 }
